pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "promptlab"
version = "0.2.0"
description = "A lightweight CLI tool for testing, comparing, and iterating on LLM prompts across models"
authors = [
    {name = "Clarity", email = "clarity.ai.bot@gmail.com"},
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "click>=8.1.0",
    "rich>=13.0.0",
    "litellm>=1.0.0",
    "pyyaml>=6.0",
    "aiohttp>=3.8.0",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
promptlab = "promptlab.cli:main"

[project.urls]
Homepage = "https://github.com/clarity-0x/promptlab"
Repository = "https://github.com/clarity-0x/promptlab.git"
Issues = "https://github.com/clarity-0x/promptlab/issues"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
line-length = 88
target-version = "py39"